requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
llama-index-vector-stores-azureaisearch
pypdf
azure-search-documents
azure-core
streamlit
httpx
langchain
langchain-openai
msal
requests
python-multipart